AMLS-4888 | Fleshed out test coverage
